Move impl to respective hooks
